win: fix invalid print format for off_t variables

+++ src/test/blk_rw/blk_rw.c
@@ -123,33 +123,33 @@ main(int argc, char *argv[])
 		switch (argv[arg][0]) {
 		case 'r':
 			if (pmemblk_read(handle, buf, lba) < 0)
-				UT_OUT("!read      lba %jd", lba);
+				UT_OUT("!read      lba %zu", (size_t)lba);
 			else
-				UT_OUT("read      lba %jd: %s", lba,
-						ident(buf));
+				UT_OUT("read      lba %zu: %s", (size_t)lba,
+					ident(buf));
 			break;
 
 		case 'w':
 			construct(buf);
 			if (pmemblk_write(handle, buf, lba) < 0)
-				UT_OUT("!write     lba %jd", lba);
+				UT_OUT("!write     lba %zu", (size_t)lba);
 			else
-				UT_OUT("write     lba %jd: %s", lba,
-						ident(buf));
+				UT_OUT("write     lba %zu: %s", (size_t)lba,
+					ident(buf));
 			break;
 
 		case 'z':
 			if (pmemblk_set_zero(handle, lba) < 0)
-				UT_OUT("!set_zero  lba %jd", lba);
+				UT_OUT("!set_zero  lba %zu", (size_t)lba);
 			else
-				UT_OUT("set_zero  lba %jd", lba);
+				UT_OUT("set_zero  lba %zu", (size_t)lba);
 			break;
 
 		case 'e':
 			if (pmemblk_set_error(handle, lba) < 0)
-				UT_OUT("!set_error lba %jd", lba);
+				UT_OUT("!set_error lba %zu", (size_t)lba);
 			else
-				UT_OUT("set_error lba %jd", lba);
+				UT_OUT("set_error lba %zu", (size_t)lba);
 			break;
 		}
 	}
